+++ src/components/game-progress.vue
