+++ src/components/game-score.vue
