+++ src/components/game-summary.vue
