+++ src/components/score-footer.vue
